Started work on Current_user library.
Current_user is being developed for the purpose of defining a user input
into our searchbar, this way we can maintain relevant data within the
object, and use the necesary functions to gather the data we need on the
user.
Halfway done implementing functions incharge of getting data from the
RIOT API, even if some seem unviable for specific users, they will still
be implemented, for possible future use.
Forget Home_Model, early implementation of Current_user, but did not
satify requirements or fit as a true Model element.